requirements.txt
# Core Data Science
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine Learning Models
xgboost>=2.0.0
lightgbm>=4.0.0

# Feature Store & Database
pymongo>=4.6.0
dnspython>=2.4.0

# API & Data Fetching
requests>=2.31.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Explainability
shap>=0.43.0

# Streamlit Dashboard
streamlit>=1.29.0

# Utilities
joblib>=1.3.0